Bug: fix image close button ++ (#56)
### Changes done in this pr:
- Made the fullscreen image overlay its own component and it is now
rendering outside the main container, so that it can be placed in front
over the headercontainer. This also solved the problem of the cross out
button not being visible
- Changed the styling of the preview images in the input field
- <img width="1132" height="275" alt="image"
src="https://github.com/user-attachments/assets/45c8217c-8b60-4b5b-af0f-f8acea8283d6"
/>

- Added fullscreen functionality to preview images as well.
- Made the email link in the helper text work again by moving helpertext
to ChatInput.tsx and make the buttons class smaller, so that it is not
in front of the helper text.
- Fixed the logo link to match brukerstotte vs servicedesk
- Added margin-bottom to the chatloader in order for it to still be
visible after the page has scrolled.

### Something I tried
<img width="1134" height="269" alt="image"
src="https://github.com/user-attachments/assets/2a244adf-27d2-4f6a-8855-5a4cfd1579bc"
/>
I tried adding a checkerboard background to highlight transparent
images, but felt that it was a bit ugly. I am however open to adding it
back 😅. Could also potentially add it to the fullscreen overlay

### To do next in order:
- Translate all text that is not translated yet 
- Switch to brukerstotte mail when on brukerstotte
- Make the image error message a popup
- Fix the color of the text in the dropdown menues
- Make the logo in the ChatbotPage a bit smaller